fix: scope Chromium EGL workaround to HDR

+++ modules/home-manager/programs/proton-pass.nix
@@ -34,20 +34,26 @@
   pkgs,
   lib,
   features,
+  displays,
   ...
 }:
 
 let
-  protonPass = pkgs.symlinkJoin {
-    name = "proton-pass-electron-hdr-sdr";
-    paths = [ pkgs.proton-pass ];
-    nativeBuildInputs = [ pkgs.makeWrapper ];
-    postBuild = ''
-      rm -f "$out/bin/proton-pass"
-      makeWrapper "${pkgs.proton-pass}/bin/proton-pass" "$out/bin/proton-pass" \
-        --add-flags "--use-gl=egl"
-    '';
-  };
+  hasHDR = lib.any (monitor: monitor.hdr) displays.monitors;
+  protonPass =
+    if hasHDR then
+      pkgs.symlinkJoin {
+        name = "proton-pass-electron-hdr-sdr";
+        paths = [ pkgs.proton-pass ];
+        nativeBuildInputs = [ pkgs.makeWrapper ];
+        postBuild = ''
+          rm -f "$out/bin/proton-pass"
+          makeWrapper "${pkgs.proton-pass}/bin/proton-pass" "$out/bin/proton-pass" \
+            --add-flags "--use-gl=egl"
+        '';
+      }
+    else
+      pkgs.proton-pass;
 in
 {
   #===========================
